In Class, Added some stub thumnail for video handling code.
I am going to handle the video thumbnails by just using the fileManager as normal,
and I'm going to have a function that detects what type of file it is and sets the ImageView accordingly.

Then replace all the existing image set calls with this function, and then it should work just fine.